web: combine scans blueprint into views / api blueprint
so headers apply to all routes
