+++ depobs/website/views.py
@@ -1,24 +1,20 @@
-from datetime import datetime
+from datetime import datetime, timedelta
 import logging
-from typing import Dict, List, Optional
+from typing import Dict, List, Optional, Tuple
 
 from flask import (
-    abort,
     Blueprint,
-    Response,
+    jsonify,
     redirect,
     request,
-    send_from_directory,
     url_for,
 )
 from werkzeug.exceptions import BadRequest, NotFound
 
 from depobs.database import models
-from depobs.website.scans import (
-    validate_npm_package_version_query_params,
-    validate_scored_after_ts_query_param,
-)
 from depobs.website.celery_tasks import get_celery_tasks
+import depobs.worker.validators as validators
+
 
 log = logging.getLogger(__name__)
 
@@ -276,3 +272,81 @@ def add_standard_headers_to_static_routes(response):
 @api.route("/")
 def index_page():
     return send_from_directory("static/", "index.html")
+
+
+def validate_scored_after_ts_query_param() -> datetime:
+    param_values = request.args.getlist("scored_after_ts", int)
+    if len(param_values) > 1:
+        raise BadRequest(description="only one scored_after_ts param supported")
+    param_value = param_values[0] if len(param_values) else None
+    # utcfromtimestamp might raise https://docs.python.org/3/library/exceptions.html#OverflowError
+    return (
+        datetime.utcfromtimestamp(param_value)
+        if param_value
+        else (datetime.now() - timedelta(days=90))
+    )
+
+
+def validate_npm_package_version_query_params() -> Tuple[str, str, str]:
+    package_names = request.args.getlist("package_name", str)
+    package_versions = request.args.getlist("package_version", str)
+    package_managers = request.args.getlist("package_manager", str)
+    if len(package_names) != 1:
+        raise BadRequest(description="only one package name supported")
+    if len(package_versions) > 1:
+        raise BadRequest(description="only zero or one package version supported")
+    if len(package_managers) > 1:
+        raise BadRequest(description="only one package manager supported")
+
+    package_name = package_names[0]
+    package_version = (
+        package_versions[0]
+        if (len(package_versions) and package_versions[0] != "")
+        else None
+    )
+    package_manager = package_managers[0] if len(package_managers) else "npm"
+
+    package_name_validation_error = validators.get_npm_package_name_validation_error(
+        package_name
+    )
+    if package_name_validation_error is not None:
+        raise BadRequest(description=str(package_name_validation_error))
+
+    if package_version:
+        package_version_validation_error = validators.get_npm_package_version_validation_error(
+            package_version
+        )
+        if package_version_validation_error is not None:
+            raise BadRequest(description=str(package_version_validation_error))
+
+    if package_manager != "npm":
+        raise BadRequest(description="only the package manager 'npm' supported")
+
+    return package_name, package_version, package_manager
+
+
+@api.route("/scan", methods=["POST"])
+def scan():
+    package_name, package_version, _ = validate_npm_package_version_query_params()
+    result: celery.result.AsyncResult = get_celery_tasks().scan_npm_package.delay(
+        package_name, package_version
+    )
+    return dict(task_id=result.id)
+
+
+@api.route("/build_report_tree", methods=["POST"])
+def build_report_tree():
+    package_name, package_version, _ = validate_npm_package_version_query_params()
+    result: celery.result.AsyncResult = get_celery_tasks().build_report_tree.delay(
+        (package_name, package_version)
+    )
+    return dict(task_id=result.id)
+
+
+@api.route("/scan_then_build_report_tree", methods=["POST"])
+def scan_npm_package_then_build_report_tree():
+    package_name, package_version, _ = validate_npm_package_version_query_params()
+    result: celery.result.AsyncResult = get_celery_tasks().scan_npm_package_then_build_report_tree.delay(
+        package_name, package_version
+    )
+    return dict(task_id=result.id)
